+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and commits should be formatted using [Conventional Commits](https://www.convent
 
 ### Added
 
-- Loader: Skip dump file when no incremental changes are detected
+- Loader: Skip dump file when no incremental changes are detected by @stronk7 ([57106dd](https://github.com/moodlehq/wiki-rag/commit/57106dddf4241db64a105d5269d93d31da38a1ea))
+- Indexer: Skip re-indexing when dump was already indexed
 
 ## [0.13.1] - 2026-03-23